speech-analyzer-webservice/requirements.txt
wheel
gunicorn
flask
python-dotenv
requests
boto3